Add files via upload
updated on 4-9-16

+++ shcart9/src/main/java/com/niit/shoppingcart/dao/CartDAO.java
@@ -10,5 +10,5 @@ public interface CartDAO {
 	public Cart get(String id);
 	public void saveOrUpdate(Cart Cart);
 	public String delete(String id);
-	public int getTotalAmount(String id);
+	public Long getTotalAmount(String id);
 }
